requirements.txt
numpy~=1.18.5
opencv-python~=4.2.0.34
addict~=2.2.1
six~=1.15.0
torchvision>=0.3.0
terminaltables~=3.1.0
pillow~=7.1.2

tqdm~=4.46.0
matplotlib~=3.2.2